+++ crates/flz/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "maili-flz"
+description = "Tiny FastLZ compression library"
+
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+exclude.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+
+[dev-dependencies]
+revm.workspace = true
+rstest.workspace = true
+alloy-sol-types.workspace = true
