+++ crates/flz/README.md
@@ -1,0 +1,18 @@
+## `maili-flz`
+
+<a href="https://github.com/op-rs/maili/actions/workflows/ci.yml"><img src="https://github.com/op-rs/maili/actions/workflows/ci.yml/badge.svg?label=ci" alt="CI"></a>
+<a href="https://crates.io/crates/maili-flz"><img src="https://img.shields.io/crates/v/maili-flz.svg" alt="maili-flz crate"></a>
+<a href="https://github.com/op-rs/maili/blob/main/LICENSE-MIT"><img src="https://img.shields.io/badge/License-MIT-d1d1f6.svg?label=license&labelColor=2a2f35" alt="MIT License"></a>
+<a href="https://github.com/op-rs/maili/blob/main/LICENSE-APACHE"><img src="https://img.shields.io/badge/License-APACHE-d1d1f6.svg?label=license&labelColor=2a2f35" alt="Apache License"></a>
+<a href="https://op-rs.github.io/maili"><img src="https://img.shields.io/badge/Book-854a15?logo=mdBook&labelColor=2a2f35" alt="Book"></a>
+
+Tiny crate containing FastLZ compression length utilities.
+
+### Provenance
+
+This code was initially duplicated across a few places including
+[maili-protocol][maili-protocol], [revm][revm], and [reth][reth].
+
+[revm]: https://github.com/bluealloy/revm
+[reth]: https://github.com/paradigmxyz/reth
+[maili-protocol]: https://github.com/op-rs/maili-protocol
