+++ crates/protocol/src/lib.rs
@@ -48,7 +48,7 @@ mod iter;
 pub use iter::FrameIter;
 
 mod utils;
-pub use utils::{read_tx_data, starts_with_2718_deposit, to_system_config};
+pub use utils::{flz_compress_len, read_tx_data, starts_with_2718_deposit, to_system_config};
 
 mod channel_out;
 pub use channel_out::{ChannelOut, ChannelOutError};
